backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
openai==1.58.1
deep-translator==1.11.4
gTTS==2.5.4
python-dotenv==1.0.1
